+++ src/Components/SelectableCard/SelectableCard.tsx
@@ -1,5 +1,5 @@
 // Libraries
-import React, {forwardRef, MouseEvent} from 'react'
+import React, {forwardRef, MouseEvent, KeyboardEvent} from 'react'
 import classnames from 'classnames'
 
 // Components
@@ -22,7 +22,9 @@ export interface SelectableCardProps extends StandardFunctionProps {
   /** Text label */
   label: string
   /** Useful for toggling selected state */
-  onClick: (id?: string) => void
+  onClick: (id?: string, e?: MouseEvent<SelectableCardRef>) => void
+  /** Useful for toggling selected state */
+  onKeyDown?: (id?: string, e?: KeyboardEvent<SelectableCardRef>) => void
   /** Controls font size of the card's label */
   fontSize?: ComponentSize
   /** Controls the color of the selected border */
@@ -35,6 +37,8 @@ export interface SelectableCardProps extends StandardFunctionProps {
   formName?: string
   /** Customize the icon that appears in selected state */
   icon?: IconFont
+  /** Customize the icon that appears in selected state */
+  tabIndex?: number
 }
 
 export type SelectableCardRef = HTMLDivElement
@@ -56,7 +60,9 @@ export const SelectableCard = forwardRef<
       selected = false,
       disabled = false,
       formName,
+      tabIndex,
       children,
+      onKeyDown,
       className,
     },
     ref
@@ -69,11 +75,15 @@ export const SelectableCard = forwardRef<
       [`${className}`]: className,
     })
 
-    const handleClick = (e: MouseEvent<HTMLDivElement>): void => {
-      e.preventDefault()
-
+    const handleClick = (e: MouseEvent<SelectableCardRef>): void => {
       if (!disabled) {
-        onClick(id)
+        onClick(id, e)
+      }
+    }
+
+    const handleKeyDown = (e: KeyboardEvent<SelectableCardRef>): void => {
+      if (!disabled && onKeyDown) {
+        onKeyDown(id, e)
       }
     }
 
@@ -85,8 +95,10 @@ export const SelectableCard = forwardRef<
         ref={ref}
         style={style}
         onClick={handleClick}
+        onKeyDown={handleKeyDown}
         className={selectableCardClass}
         data-testid={testID}
+        tabIndex={disabled ? undefined : tabIndex}
       >
         <label className="cf-selectable-card--label" htmlFor={id}>
           {label}
@@ -105,6 +117,7 @@ export const SelectableCard = forwardRef<
             value={label}
             defaultChecked={selected}
             disabled={disabled}
+            tabIndex={-1}
           />
         </div>
       </div>